requirements/base.txt
django-extensions==2.1.4
Django==3.1.12
django-environ==0.4.5
django-model-utils==4.1.1
django-mptt==0.12.0
django-redis==5.0.0
djangorestframework==3.11.2
mongoengine==0.22.1
requests==2.23.0
django-filter==2.4.0
transitions
django-mptt
